Add files via upload
The Buzzer Module is an electronic component that emits sound when powered. In this project, it serves as an auditory alert system, notifying the user when an object is detected within a specified range by the ultrasonic sensor. The module typically has two pins: one for power (VCC) and one for ground (GND). By connecting the VCC pin to the Arduino's 5V pin and the GND pin to the Arduino's GND, the buzzer can be activated through the Arduino's digital output pins.